requirements.txt
firebase-admin==7.1.0
google-auth==2.48.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
